+++ Asteroids/src/com/neet/managers/ScoreComparator.java
@@ -5,8 +5,8 @@ import java.util.Comparator;
 public class ScoreComparator implements Comparator<Score>{
 
 	public int compare(Score score1, Score score2) {
-		int sc1 = score1.getScore();
-		int sc2 = score2.getScore();
+		long sc1 = score1.getScore();
+		long sc2 = score2.getScore();
 		
 		if(sc1 > sc2) {
 			return - 1;
